requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.2.4
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-core==0.3.83
langchain-groq==0.3.8
langchain-ollama==0.3.10
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.16.0
networkx==3.2.1
numpy==2.0.2
ollama==0.6.1
openpyxl==3.1.5
oracledb==3.4.2
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.5
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.8.0
tornado==6.5.4
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
watchdog==6.0.0
Werkzeug==3.1.5
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
